+++ libs/core/src/components/text/text.ts
@@ -6,6 +6,7 @@ import { GdsElement } from '../../gds-element'
 import { tokens } from '../../tokens.style'
 import { styleExpressionProperty } from '../../utils/decorators/style-expression-property'
 import { gdsCustomElement } from '../../utils/helpers/custom-element-scoping'
+import { defaultStyles } from './default-typography.styles'
 import textStyles from './text.style'
 
 /**
@@ -18,7 +19,7 @@ import textStyles from './text.style'
  */
 @gdsCustomElement('gds-text')
 export class GdsText extends GdsElement {
-  static styles = [tokens, textStyles]
+  static styles = [tokens, defaultStyles, textStyles]
 
   /**
    * The level of the container can be used to apply background and color styles from the corresponding level.
@@ -198,8 +199,14 @@ export class GdsText extends GdsElement {
    * @property lines
    */
   @styleExpressionProperty({
-    property: '--_lines',
     valueTemplate: (v) => v,
+    styleTemplate: (_prop, values) => {
+      return `overflow: hidden;
+      text-overflow: ellipsis;
+      display: -webkit-box;
+      -webkit-line-clamp: ${values[0]};
+      -webkit-box-orient: vertical;`
+    },
   })
   lines?: number
 
@@ -228,13 +235,7 @@ export class GdsText extends GdsElement {
   color?: string
 
   render() {
-    const TAG_ENCODE = encodeURI(this.tag)
-    const TAG = unsafeStatic(TAG_ENCODE)
-    const classes = {
-      'no-size-set': !this['font-size'],
-      'no-weight-set': !this['font-weight'],
-      'lines-set': !!this.lines,
-    }
-    return html`<${TAG} tag class=${classMap(classes)}><slot></slot></${TAG}>`
+    const TAG = unsafeStatic(encodeURI(this.tag))
+    return html`<${TAG} tag><slot></slot></${TAG}>`
   }
 }