+++ libs/react/src/lib/modal/modal.stories.tsx
@@ -38,6 +38,7 @@ export const ModalDefault: Story<ModalProps> = Template.bind({})
 ModalDefault.args = {
   type: 'default',
   header: 'Default Modal',
+  closeText: 'Custom Close Label',
   children: 'Body content',
   confirm: 'OK',
   dismiss: 'Nope',
@@ -46,7 +47,8 @@ ModalDefault.args = {
 export const SlideOutSmall: Story<ModalProps> = Template.bind({})
 SlideOutSmall.args = {
   type: 'slideout',
-  header: 'SlideOut Modal',
+  header: 'SlideOut Modal - Small',
+  closeText: 'Custom Close Label',
   children: 'Body content',
   confirm: 'OK',
   dismiss: 'Nope',
@@ -56,7 +58,8 @@ SlideOutSmall.args = {
 export const SlideOutMedium: Story<ModalProps> = Template.bind({})
 SlideOutMedium.args = {
   type: 'slideout',
-  header: 'SlideOut Modal',
+  header: 'SlideOut Modal - Medium',
+  closeText: 'Custom Close Label',
   children: 'Body content',
   confirm: 'OK',
   dismiss: 'Nope',
@@ -66,7 +69,8 @@ SlideOutMedium.args = {
 export const SlideOutLarge: Story<ModalProps> = Template.bind({})
 SlideOutLarge.args = {
   type: 'slideout',
-  header: 'SlideOut Modal',
+  header: 'SlideOut Modal - Large',
+  closeText: 'Custom Close Label',
   children: 'Body content',
   confirm: 'OK',
   dismiss: 'Nope',
@@ -77,6 +81,7 @@ export const TakeOver: Story<ModalProps> = Template.bind({})
 TakeOver.args = {
   type: 'takeover',
   header: 'TakeOver Modal',
+  closeText: 'Custom Close Label',
   children: 'Body content',
   confirm: 'OK',
   dismiss: 'Nope',