+++ libs/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.20.2
+
+### Patch Changes
+
+- 2b58b66: Modal react close button fix
+- Updated dependencies [de5d0f2]
+  - @sebgroup/green-core@1.48.1
+
 ## 3.20.1
 
 ### Patch Changes